Use a percentage instead of fraction for the remote connections percentage CLI option. (#1682)
* Factorise metrics code for KeyValueStorage database

- introduce `MonitorableKeyValueStorage`
- factorise code
- remove metrics instanciation in `RocksDbKeyValueStorage` and `ColumnarRocksDbKeyValueStorage`

* Rename class

* Use percentage instead of fraction

* add tests

* fix PR discussion

- introduce `PercentageConverter` and `FractionConverter` in `tech.pegasys.pantheon.cli.converter`
- use custom error message for `PercentageConversionException
`

* fix PR discussion

- return primitive types instead of objects for converters

* fix tests

+++ pantheon/src/main/java/tech/pegasys/pantheon/cli/DefaultCommandValues.java
@@ -65,7 +65,7 @@ public interface DefaultCommandValues {
   int FAST_SYNC_MAX_WAIT_TIME = 0;
   int FAST_SYNC_MIN_PEER_COUNT = 5;
   int DEFAULT_MAX_PEERS = 25;
-  double DEFAULT_FRACTION_REMOTE_WIRE_CONNECTIONS_ALLOWED =
+  float DEFAULT_FRACTION_REMOTE_WIRE_CONNECTIONS_ALLOWED =
       RlpxConfiguration.DEFAULT_FRACTION_REMOTE_CONNECTIONS_ALLOWED;
 
   static Path getDefaultPantheonDataPath(final Object command) {
